rename HIRT_Multichannel_Convolution/{ConvolveMono => mono}.hpp
